requirements.txt
asgiref==3.5.0
Django==4.0.3
django-bootstrap-modal-forms==2.2.0
sqlparse==0.4.2
